Fix: Separate components by spinner, section spinner
